Add output config to sway
This will only work on my setup. Even though I have almost everything
else configured so it applies to most setups, this seems to be
impossible here. The same applies to my gitconfig.
I am pretty sure that this will not necessarily cause any complications,
since sway seems to use a fallback output anyways if none of the
configured outputs are detected.